refactor: extract gateway pairing api surface
